Support code insertion in untitled and notebook file.
